… tests/test_regex_parsing.py