Various performance commits. 

```
A 
0 errors across 0 files from annotations. 150 errors in the store
group_by_filename_and_update took 35ms [D]  --> 4ms

0 errors across 0 files from flake8. 150 errors in the store
group_by_filename_and_update took 75ms [D]  --> 23ms

8 errors across 5 files from mypy. 150 errors in the store
group_by_filename_and_update took 113ms [D] --> 16ms
```
```
B
0 errors across 0 files from annotations. 150 errors in the store
group_by_filename_and_update took 166ms [D] --> 4ms

0 errors across 0 files from flake8. 150 errors in the store
group_by_filename_and_update took 127ms [D] --> 25ms

58 errors across 5 files from mypy. 150 errors in the store
group_by_filename_and_update took 292ms [D] --> 23ms
```
```
C
0 errors across 0 files from annotations. 150 errors in the store
group_by_filename_and_update took 83ms [D]  --> 8ms

0 errors across 0 files from flake8. 150 errors in the store
group_by_filename_and_update took 46ms [D]  --> 7ms

70 errors across 29 files from mypy. 150 errors in the store
group_by_filename_and_update took 975ms [D] --> 42ms
```

Must be read commit by commit. 

Solves #1623 for free in a pragmatic way: if the results change it will open the panel again. Otherwise the panel stays closed.

`settings` are now generally cached in memory. Greatly improves `settings.has_changed` which was very cheap and idiosyncratic and now just works as anyone would expect. 

Introduces new event `lint_result_changed` which otherwise is the same as our known 'lint_result' with a cache in front. Notable: Since we usually rely on our settings as a global, t.i. we're never pure anywhere probably, we use a settings 'change-count' as additional cache salt.



Fixes #1623
Fixes #1662 
Fixes #1671