@kaste 2019-10-05
I don't like the complicated throttler in panel_view so I reverted e803a44 and the previous throttler 2cd32b7. 

Now, 5284df0 and the new tip b76c6b0 have the same performance. I think all commits are safe, easy to grasp, except e30f67a is (maybe?) tricky. 

With the performance now we basically hit the wall (with warm cache), in that the low ms 5, 6, 8 are spent in `group_by_filename_and_update` itself and not in the event listeners/view anymore.