@braver 2019-10-05
Nifty. Are you ready for a review here or want to do more work first?