@kaste 2019-10-05
This is good enough for review. Sure I can beautify and add some comments and better commit messages ... but I think this works and should be safe.