@braver 2019-10-05
Huh, I was under the impression that accessing settings via the ST API was essentially free. Maybe there is a trick to it, I'm not sure, it's been a while. Maybe `load_settings` is not free, whereas get any particular value will get it from memory directly. Not doing a `load_settings` if you don't need to sounds like a good idea anyway.

So, yeah, looks good indeed. I like how you improved the panel view, it's actually a lot simpler to read whereas e803a44 was a bit much. 

I'll run with it for a few days, see if any errors turn up.