@kaste 2019-10-11
I maybe go for the extra mile and open separate PRs here. Looks like these are good improvements/patches with clear boundaries.

Also looks like that while we have on_lint_result which fires unconditionally we could just invent lint_result_changed. Usually a view doesn't need to update for the same result set. (This wasn't important for single file linters bc it is at least likely that the report changes for the file you're editing, but for multi file linters it is not so. You edit one file and the other ten files reported probably stay the same.) Don't need to do this immediately but it sounds reasonable.