@kaste 2019-10-15
No separate PR but I actually introduce 'lint_result_changed' bc that's really nicer and makes sense. 

Please look at `highlights_view` bc I compute the gutter region slightly different. I don't see a difference visually.  Maybe you know why we computed a 'line-region' for these icons.  On my Sublime, it works with the 'normal' region just as well. 

LGTM