@braver 2019-10-15
Maybe something about multiple errors on a line, or otherwise overlapping regions?