@kaste 2019-10-15
Good idea! The line comes from SL3 where the actual region was used to filter overlapping errors, but we just use the reported line for that now.