@braver 2019-10-15
I’ll try and put it through the ringer later today. Other than that the branch has been really solid for me.