@braver 2019-10-15
Ok, it looked good enough for what I tried now, but I need to try something more complicated with eslint. Will try to break it tomorrow and will report back.