@braver 2019-10-16
Ok, most of this looks and behaves well. I can't remember how to trigger overlapping multi-line errors. I seem to remember that happening with TSLint sometimes, but I haven't done any TypeScript for some time now and don't have any projects lying around. 

That said, there is one problem: the tooltips on the gutter no longer work. On the regions themselves they're fine but this doesn't work anymore (screenshot taken on master while hovering over the line number):

<img width="453" alt="Screen Shot 2019-10-16 at 22 20 44" src="https://user-images.githubusercontent.com/2543659/66956016-dd7e9600-f063-11e9-94e7-485777771d57.png">