@kaste 2019-10-17
Workflow question: How do you checkout after I force-push?