@braver 2019-10-18
Well usually I just pull and then notice it doesn’t fast-forward. So then I switch to master, remove the branch, and checkout the branch again.